node usize

+++ engine/src/xetex_ini.rs
@@ -2901,7 +2901,7 @@ unsafe fn store_fmt_file() {
     p = avail;
     while !p.is_texnull() {
         dyn_used -= 1;
-        p = *LLIST_link(p as isize)
+        p = *LLIST_link(p as usize)
     }
     fmt_out.dump_one(var_used as i32);
     fmt_out.dump_one(dyn_used as i32);
@@ -4904,7 +4904,7 @@ unsafe fn initialize_primitives() {
     primitive(b"vfilneg", VSKIP, FIL_NEG_CODE);
     primitive(b"mskip", MSKIP, MSKIP_CODE);
 
-    primitive(b"kern", KERN, EXPLICIT);
+    primitive(b"kern", KERN, EXPLICIT as i32);
     primitive(b"mkern", MKERN, MU_GLUE as i32);
     primitive(b"moveleft", HMOVE, 1);
     primitive(b"moveright", HMOVE, 0);
